README.md
Overview

AracKiralama is a Python-based project designed to manage and streamline car rental operations. It includes a graphical user interface (GUI) created with PyQt6, enabling users to interact with the system efficiently. It was developed with Python 3.12, QT 15.0.0 and MySQL 8.0. And the application is in TURKISH.

Features

 *User-friendly interface for car rental management.

 *Multiple UI components for various functionalities.

 *Modular design with separate files for UI and application logic.

Project Structure

 *anapencere.py: Main application logic and functionality.

 *.ui: Design files for the graphical user interface (created using Qt Designer).

 ui_*.py: Auto-generated Python files from the .ui design files.

 *requirements.txt: List of required Python packages to run the project.

 *.qtcreator: Project configuration files for Qt Creator IDE.

 *__pycache__: Directory containing cached Python bytecode files.

Prerequisites

To run the project, ensure the following are installed:

 *Python 3.11 or above

 *Required Python libraries (specified in requirements.txt)